globals consolidation WIP
