+++ contract-shared-headers/external_types.hpp
