+++ contracts/daccustodian/voting.cpp
@@ -5,11 +5,11 @@ ACTION daccustodian::votecust(const name &voter, const vector<name> &newvotes, c
     check(dac_id == "testa"_n, "Voting is not yet enabled in the Planet DAOs.");
 #endif
     candidates_table registered_candidates(_self, dac_id.value);
-    contr_config     configs = contr_config::get_current_configs(_self, dac_id);
+    const auto       globals = dacglobals::current(get_self(), dac_id);
 
     require_auth(voter);
     assertValidMember(voter, dac_id);
-    check(newvotes.size() <= configs.maxvotes,
+    check(newvotes.size() <= globals.get_maxvotes(),
         "ERR::VOTECUST_MAX_VOTES_EXCEEDED::Max number of allowed votes was exceeded.");
     std::set<name> dupSet{};
     assertValidMembers(newvotes, dac_id);
